Django/todolist/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.6
djangorestframework==3.14.0
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
